more refactoring, delete now unused files/functions
