+++ src/arcsf/eval/utils.py
@@ -1,16 +1,8 @@
 import warnings
 
-import matplotlib.pyplot as plt
 import numpy as np
 import torch
-import transformers
-from accelerate import Accelerator
-from torch.utils.data import DataLoader
-from tqdm import tqdm
 from transformers import DataCollatorWithPadding, PreTrainedTokenizer
-
-from arcsf.data.data_module import EvalQADataset, EvaluateDataCollator
-from arcsf.eval.metrics import ecdf, eval_rouge_recall, get_loss, ks_test, truth_ratio
 
 
 def check_nans(array: np.ndarray | torch.Tensor, name: str = "") -> None:
@@ -23,29 +15,6 @@ def check_nans(array: np.ndarray | torch.Tensor, name: str = "") -> None:
     elif isinstance(array, torch.Tensor):
         if torch.isnan(array).any():
             warnings.warn(message)
-
-
-def combine_dicts(
-    forget_dict: dict[str, np.ndarray | torch.Tensor],
-    retain_dict: dict[str, np.ndarray | torch.Tensor],
-) -> dict[str, torch.Tensor]:
-    """
-    Combines the evaluation dictionaries from the forget and retain data to return only
-    the values used for the aggregation.
-
-    Args:
-        forget_dict: Evaluation metrics from the forget data
-        retain_dict: Evaluation metrics from the retain data
-
-    Returns:
-        Dictionary containing the retain truth ratios, forget truth ratios, and forget
-        ROUGE scores.
-    """
-    return {
-        "retain_tr": retain_dict["truth_ratios"],
-        "forget_tr": forget_dict["truth_ratios"],
-        "rouge_scores": retain_dict["rougeL_recall"],
-    }
 
 
 def first_idx(
@@ -126,214 +95,3 @@ def extract_qa_for_generate(
     questions["attention_mask"] = questions["attention_mask"][:, n_trim_pad:]
 
     return questions, answers
-
-
-def all_eval(
-    model: transformers.PreTrainedModel,
-    dataset: EvalQADataset,
-    batch_size: int,
-    tokenizer: transformers.PreTrainedTokenizer,
-    n_print: int = 0,
-    accelerator: Accelerator = Accelerator(),
-    **generate_kwargs: dict,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Performs evaluation of the selected model over selected data. Returns the
-    input-wise losses on target answers for the ground truth and perturbed answers.
-
-    Args:
-        model : Transformers model used to perform evaluation on
-        dataset : Dataset to perform evaluation on
-        batch_size : batch size for dataloader
-        device : Pytorch device on which to perform computation
-        n_print : number of samples to print the generated answers for
-        accelerator : Accelerator object to handle moving data etc. to device
-
-    Returns:
-        output_dict : output dictionary with the values for analysis
-    """
-    n_perturbed = dataset.n_perturbed
-    tokenizer.padding_side = "left"
-    eval_collate_fn = EvaluateDataCollator(tokenizer=tokenizer)
-    data_loader = DataLoader(
-        dataset,
-        batch_size=batch_size,
-        shuffle=False,
-        collate_fn=eval_collate_fn,
-    )
-
-    # Make data loader that returns batches of separated questions and ground truth
-    # answers (so this can then be passed to accelerate to handle moving data to device)
-    # batch[0] below selects the ground truth q/a pair (rather than perturbed answers)
-    # batch size 1 as batch already created in previous data loader
-    qa = [extract_qa_for_generate(batch[0], tokenizer) for batch in data_loader]
-    qa_loader = DataLoader(qa, batch_size=1, collate_fn=lambda x: x[0], shuffle=False)
-
-    # handle moving model and data loaders to device
-    model, data_loader, qa_loader = accelerator.prepare(model, data_loader, qa_loader)
-
-    dataset_len = len(dataset)
-    output_dict = {
-        "all_losses": torch.zeros((dataset_len, n_perturbed + 1), dtype=torch.float64),
-        "truth_ratios": torch.zeros(dataset_len),
-        "rougeL_recall": torch.zeros(dataset_len),
-        "rouge1_recall": torch.zeros(dataset_len),
-    }
-
-    # =========
-    # Metrics on logits of ground truth and perturbed answers
-    # =========
-    for batch_idx, batch in enumerate(
-        tqdm(data_loader, desc="GT and Perturbed Logits")
-    ):
-        gt_batch = batch[0]
-        pt_batches = batch[1:]
-        batch_start_index = batch_idx * batch_size
-        batch_end_index = batch_start_index + batch_size
-
-        # Ground truth answers
-        with torch.no_grad():
-            # DO pass position_ids into this method -> see collator for info
-            gt_outputs = model(**gt_batch)
-
-        gt_loss = get_loss(gt_outputs.logits, gt_batch["labels"])
-        output_dict["all_losses"][batch_start_index:batch_end_index, 0] = gt_loss.cpu()
-
-        # Perturbed answers
-        for perturbed_index in range(n_perturbed):
-            pt_batch = pt_batches[perturbed_index]
-            with torch.no_grad():
-                p_output = model(**pt_batch)
-            output_dict["all_losses"][
-                batch_start_index:batch_end_index, perturbed_index + 1
-            ] = get_loss(p_output.logits, pt_batch["labels"]).cpu()
-
-    # =========
-    # Metrics on generated answers vs. actual ground truth answers
-    # =========
-    for batch_idx, qa in enumerate(tqdm(qa_loader, desc="Generate")):
-        batch_start_index = batch_idx * batch_size
-        questions, answers = qa
-
-        with torch.no_grad():
-            # DO NOT pass position_ids into this method -> see collator for info
-            gen_outputs = model.generate(
-                input_ids=questions["input_ids"],
-                attention_mask=questions["attention_mask"],
-                pad_token_id=tokenizer.pad_token_id,
-                **generate_kwargs,
-            )
-
-        target_answers = tokenizer.batch_decode(
-            answers["input_ids"], skip_special_tokens=True
-        )
-        generated_answers = [
-            tokenizer.decode(
-                gen_a[len(q) :],  # only want the tokens for the answers
-                skip_special_tokens=True,
-            )
-            for q, gen_a in zip(questions["input_ids"], gen_outputs)
-        ]
-
-        if batch_idx * batch_size < n_print:
-            n_print_batch = min(batch_size, n_print - batch_idx * batch_size)
-            question_text = tokenizer.batch_decode(
-                questions["input_ids"][:n_print_batch]
-            )
-            for q_text, gen_text, target_text in zip(
-                question_text,
-                generated_answers[:n_print_batch],
-                target_answers[:n_print_batch],
-            ):
-                print(f"\nQuestion: {q_text}\n")
-                print(f"Target: {target_text}\n")
-                print(f"Generated: {gen_text}")
-                print("-" * 15)
-
-        for rouge_idx, (generated_text, target_text) in enumerate(
-            zip(generated_answers, target_answers)
-        ):
-            rouge_result = eval_rouge_recall(
-                gen_output=generated_text, ground_truth=target_text
-            )
-            output_dict["rougeL_recall"][batch_start_index + rouge_idx] = rouge_result[
-                "rougeL_recall"
-            ]
-            output_dict["rouge1_recall"][batch_start_index + rouge_idx] = rouge_result[
-                "rouge1_recall"
-            ]
-
-    # calculate truth_ratio and return them along with losses
-    output_dict["truth_ratios"] = truth_ratio(output_dict["all_losses"])
-
-    return output_dict
-
-
-def get_analysis_values(
-    model_dir: str,
-) -> dict[str, np.ndarray | torch.Tensor]:
-    """
-    Gets the values for analysis given a model directory.
-    Args:
-        model_dir : file path to where model is stored with evaluation output
-
-    Returns:
-        vals : dictionary containing the values to be used in analysis
-    """
-    vals = dict()
-    # losses are of shape (n_samples x n_perturbed)
-    vals["forget_losses"] = np.loadtxt(model_dir + "/eval/forget/all_losses.txt")
-    vals["retain_losses"] = np.loadtxt(model_dir + "/eval/retain/all_losses.txt")
-    vals["rouge_scores"] = np.loadtxt(model_dir + "/eval/retain/rougeL_scores.txt")
-    # we re-calculate the truth ratio, since torch calculated many as NaNs
-    # truth ratios are of shape (n_samples)
-    vals["forget_tr"] = truth_ratio(torch.tensor(vals["forget_losses"]))
-    vals["retain_tr"] = torch.clamp(
-        1 - truth_ratio(torch.tensor(vals["retain_losses"])),
-        min=0,
-    )
-    for key in vals:
-        check_nans(vals[key], key)
-    return vals
-
-
-def plot_cdf(
-    base_vals: dict,
-    test_vals: dict,
-    save_path: str,
-    exp_name: str,
-    split: str = "retain",
-) -> None:
-    """
-    Function to plot and save CDF functions for model comparison
-
-    Args:
-        base_vals : values for the base model to be compared against
-        test_vals : values for the test model to be compared
-        save_path : relative path where the model is stored
-        exp_name : the name of the experiment for naming purposes
-        split (optional): the split to be used. Defaults to "retain".
-    """
-    base_data, base_y_values = ecdf(base_vals[f"{split}_tr"])
-    test_data, test_y_values = ecdf(test_vals[f"{split}_tr"])
-
-    p_val = ks_test(test_vals[f"{split}_tr"], base_vals[f"{split}_tr"])
-
-    plt.title(f"{split} data CDF")
-    plt.plot(base_data, base_y_values, label="base-model")
-    plt.plot(test_data, test_y_values, label=f"forget-model {exp_name}")
-    plt.annotate(f"Model Utility {np.log(p_val)}", (0.9, 0.1))
-    plt.legend()
-    plt.xlim(0, 1)
-    plt.savefig(save_path + f"/eval/analysis/tr-cdf-{split}-{exp_name}-.pdf")
-    plt.close()
-
-
-def plot_scatter(dict: dict[float], **plot_kwargs) -> None:
-    """
-    Plot of the data from a given model/experiment
-
-    Args:
-        dict : dictionary containing the aggrevated evaluation metrics of the model.
-        plot_kwargs : key word arguments for the `matplotlib.pylot.scatter` function.
-    """
-    plt.scatter(dict["model_utility"], dict["forget_quality_1"], **plot_kwargs)
